backend/requirements.txt
fastapi
Pillow
uvicorn